requirements.txt
# Core
openai>=1.0.0
neo4j>=5.0.0
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# PDF processing
PyMuPDF>=1.23.0

# API clients
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# YAML processing
pyyaml>=6.0.0
